refactor(statements): group statement classes within directory + create base class
