+++ src/monopoly/statements/debit_statement.py
@@ -7,12 +7,13 @@ import fitz
 from monopoly.config import DebitStatementConfig
 from monopoly.constants import AccountType, StatementFields
 from monopoly.pdf import PdfPage
-from monopoly.statement import Statement, Transaction
+
+from .base import BaseStatement, Transaction
 
 logger = logging.getLogger(__name__)
 
 
-class DebitStatement(Statement):
+class DebitStatement(BaseStatement):
     """
     A dataclass representation of a debit statement
     """
