+++ tests/integration/banks/test_banks_credit.py
@@ -7,9 +7,9 @@ from pandas.testing import assert_frame_equal
 from test_utils.skip import skip_if_encrypted
 
 from monopoly.constants import StatementFields
-from monopoly.credit_statement import CreditStatement
 from monopoly.processors import Citibank, Dbs, Hsbc, Ocbc, StandardChartered
 from monopoly.processors.base import ProcessorBase
+from monopoly.statements import CreditStatement
 
 
 @skip_if_encrypted
