+++ tests/integration/banks/test_date_handling.py
@@ -4,10 +4,10 @@ import pandas as pd
 from pandas.testing import assert_frame_equal
 
 from monopoly.processors import Ocbc
-from monopoly.statement import Statement, Transaction
+from monopoly.statements import BaseStatement, Transaction
 
 
-def test_transform_cross_year(ocbc: Ocbc, statement: Statement):
+def test_transform_cross_year(ocbc: Ocbc, statement: BaseStatement):
     raw_df = pd.DataFrame(
         [
             Transaction("12/01", "FAIRPRICE FINEST", "18.49"),
@@ -28,7 +28,7 @@ def test_transform_cross_year(ocbc: Ocbc, statement: Statement):
     assert_frame_equal(transformed_df, expected_data)
 
 
-def test_transform_within_year(ocbc: Ocbc, statement: Statement):
+def test_transform_within_year(ocbc: Ocbc, statement: BaseStatement):
     raw_df = pd.DataFrame(
         [
             Transaction("12/06", "FAIRPRICE FINEST", "18.49"),
