+++ tests/unit/test_statement_process_line.py
@@ -1,11 +1,10 @@
 import re
 
-from monopoly.constants import AccountType
 from monopoly.processors import Hsbc, Ocbc
-from monopoly.statement import Statement, Transaction
+from monopoly.statements import BaseStatement, Transaction
 
 
-def test_process_line(statement: Statement):
+def test_process_line(statement: BaseStatement):
     pattern = re.compile(Ocbc.credit_config.transaction_pattern)
     line = "19/06 YA KUN KAYA TOAST 3.20"
     lines = ["19/06 YA KUN KAYA TOAST 3.20", "20/06 FAIRPRICE FINEST 9.90"]
@@ -14,7 +13,7 @@ def test_process_line(statement: Statement):
     assert transaction == Transaction("19/06", "YA KUN KAYA TOAST", 3.2)
 
 
-def test_process_line_no_match(statement: Statement):
+def test_process_line_no_match(statement: BaseStatement):
     line = "Invalid line"
     lines = ["Invalid line"]
     idx = 0
@@ -22,7 +21,7 @@ def test_process_line_no_match(statement: Statement):
     assert transaction is None
 
 
-def test_process_line_multiline_description(statement: Statement):
+def test_process_line_multiline_description(statement: BaseStatement):
     pattern = re.compile(Hsbc.credit_config.transaction_pattern)
     statement.statement_config.multiline_transactions = True
     line = "04 Aug 02 Aug SHOPEE 3.20"
